requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.25
alembic==1.13.1
aiohttp==3.9.3
python-dotenv==1.0.1
Pillow==10.2.0
